zoek nog een fout

+++ Calculator.java
@@ -24,7 +24,7 @@ public class Calculator {
 	}
 	
 
-	public double divide(double a, double b) {
+	public double divide(int a, int b) {
 	
 	return a/b;
 	}
